Upgrade webdriverio from v7 -> v8
