+++ client-e2e/package.json
@@ -1,17 +1,18 @@
 {
-    "name": "webdriverio-tests",
-    "version": "0.1.0",
-    "private": true,
-    "devDependencies": {
-        "@wdio/cli": "^7.19.7",
-        "@wdio/local-runner": "^7.19.7",
-        "@wdio/mocha-framework": "^7.19.7",
-        "@wdio/spec-reporter": "^7.19.7",
-        "@wdio/types": "^8.1.2",
-        "ts-node": "^10.8.0",
-        "typescript": "^4.7.2"
-    },
-    "scripts": {
-        "wdio": "wdio run test/wdio.CI.conf.ts"
-    }
+  "name": "webdriverio-tests",
+  "version": "0.1.0",
+  "private": true,
+  "devDependencies": {
+    "@wdio/cli": "^8.3.10",
+    "@wdio/local-runner": "^8.3.10",
+    "@wdio/mocha-framework": "^8.3.0",
+    "@wdio/spec-reporter": "^8.3.0",
+    "@wdio/types": "^8.3.0",
+    "fs-extra": "^11.1.0",
+    "ts-node": "^10.8.0",
+    "typescript": "^4.7.2"
+  },
+  "scripts": {
+    "wdio": "wdio run test/wdio.conf.ts"
+  }
 }
